+++ src/main/java/com/github/zly2006/reden/rvc/tracking/TrackedStructure.kt
@@ -2,14 +2,22 @@ package com.github.zly2006.reden.rvc.tracking
 
 import com.github.zly2006.reden.rvc.IPlacement
 import com.github.zly2006.reden.rvc.IWritableStructure
+import com.github.zly2006.reden.rvc.RelativeCoordinate
 import com.github.zly2006.reden.rvc.tracking.network.NetworkWorker
 import com.github.zly2006.reden.utils.redenError
+import net.minecraft.block.BlockState
+import net.minecraft.nbt.NbtCompound
+import net.minecraft.util.math.BlockPos
 import net.minecraft.world.World
+import java.nio.file.Path
+import java.util.*
 
 class TrackedStructure(
     override var name: String,
     val repository: RvcRepository?,
 ) : IWritableStructure, IPlacement {
+    override val entities: MutableMap<UUID, NbtCompound>
+        get() = regions.values.flatMap { it.entities.entries }.associate { it.toPair() }.toMutableMap()
     var networkWorker: NetworkWorker? = null
     override var enabled: Boolean = true
     override val structure get() = this
@@ -33,5 +41,76 @@ class TrackedStructure(
         get() = if (networkWorker?.world?.info == placementInfo?.worldInfo)
             networkWorker?.world ?: redenError("World is not set for $name")
         else placementInfo?.worldInfo?.getWorld() ?: redenError("World is not found: $name")
+    override val origin: BlockPos
+        get() = placementInfo?.origin?.toImmutable()
+            ?: redenError("getting origin but PlacementInfo not set for $name")
 
+    fun setPlaced() {
+        require(repository != null) { "Repository is null" }
+        repository.placed = true
+        repository.placementInfo = this.placementInfo
+    }
+
+    override fun startMoving() {
+        require(repository != null) { "Repository is null" }
+        repository.placed = false
+    }
+
+    fun refreshPositionsAsync() = networkWorker?.async { refreshPositions() }
+    suspend fun refreshPositions() {
+        regions.values.forEach {
+            it.refreshPositions()
+        }
+    }
+
+    /**
+     * Remove this structure from the world, including all blocks
+     */
+    fun remove() {
+        require(repository != null) { "Repository is null" }
+        repository.placementInfo = null
+        repository.placed = false
+        clearArea()
+    }
+
+    suspend fun collectAllFromWorld() {
+        regions.values.forEach {
+            it.collectAllFromWorld()
+        }
+    }
+
+    val totalBlocks: Int get() = regions.values.sumOf { it.blocks.size }
+    val minPos: BlockPos = TODO()
+    override fun setBlockEntityData(pos: RelativeCoordinate, nbt: NbtCompound) {
+        TODO("Not yet implemented")
+    }
+
+    override fun setBlockState(pos: RelativeCoordinate, state: BlockState) {
+        TODO("Not yet implemented")
+    }
+
+    override fun getBlockEntityData(pos: RelativeCoordinate): NbtCompound? {
+        TODO("Not yet implemented")
+    }
+
+    override fun getBlockState(pos: RelativeCoordinate): BlockState {
+        TODO("Not yet implemented")
+    }
+
+    override fun getOrCreateBlockEntityData(pos: RelativeCoordinate): NbtCompound {
+        TODO("Not yet implemented")
+    }
+
+    override fun isInArea(pos: RelativeCoordinate) = regions.any { it.value.isInArea(pos) }
+    override fun createPlacement(world: World, origin: BlockPos): IPlacement {
+        val oldOffsets = regions.mapValues { it.value.placementInfo?.origin?.subtract(this.origin) }
+        regions.forEach { (k, v) ->
+            val partOrigin = oldOffsets[k]?.add(origin)
+            v.createPlacement(world, partOrigin ?: origin)
+        }
+        return this
+    }
+
+    override fun save(path: Path) = RvcFileIO.save(path, this)
+    override fun load(path: Path) = RvcFileIO.load(path, this)
 }
